move unneeded functions in hostpci to hostvf or pcidirect

+++ vancouver/include/nul/types.h
@@ -14,6 +14,6 @@ typedef unsigned long long uint64;
 typedef unsigned int       uint32;
 typedef unsigned short     uint16;
 typedef unsigned char      uint8;
-typedef unsigned int       mword;
+typedef unsigned long      mword;
 
 /* EOF */
